crdroid: Disable Deep Press touch video heatmaps
Despite redbull having a 10 MHz SPI bus for the touchscreen, transfering
heatmap data at 120 Hz is still expensive and adds latency. It also
wastes power by keeping the bus active for longer than necessary.

The deep press feature that uses heatmaps doesn't work very well anyway,
so let's just disable it.

An interesting side effect of the heatmap feature is that it increases
jitter in touch readings. Drawing curves with a single finger results in
slightly jittery lines when it is enabled, and while there is still a
little bit of jitter with it disabled, the consistency of the lines is
better overall.

While the testing and analysis for this was done on redfin, this is
still a worthwhile change for all devices that have a Deep Press input
classifier HAL and model.

Test: touch works normally on redfin
Test: latency and jitter in Touchpaint app
Change-Id: Ibc625592e62ec43e0c56fe8251825832b208db43
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>